requirements.txt
numpy
pandas
matplotlib
Pillow
scikit-learn
torch
torchvision
